requirements.txt
PyMuPDF
Pillow
pytesseract
pandas
